ft: saved links can be viewed
add a new page for links to be viewed.
At this point, the links can only be seen but it would be a great idea to add a share option or a copy
to clipboar doption.

Transitions would be nice too.

This feature also fixed the layout overflow on mobile